Extracting the core before/after functionality of each sql type to utils.
