+++ .eslintrc
@@ -4,5 +4,9 @@
   "env": {
     "node": true
   },
-  "rules": {}
+  "rules": {
+    "no-unused-vars": 0,
+    "no-undef": 0,
+    "no-console": 0
+  }
 }
