+++ src/mysql/core.js
@@ -115,82 +115,9 @@ module.exports = function(util) {
       });
   };
 
-  /**
-   * Check if the current route has a before fn defined, if so, execute it and proceed.
-   *
-   * @param {object} route
-   *   The route object.
-   * @param {object} req
-   *   The express request object.
-   * @param {object} res
-   *   The express response object.
-   *
-   * @returns {Promise}
-   */
-  var before = function before(route, req, res) {
-    debug('Before:');
-    debug(route);
-
-    // Ensure they can hook into the before handler.
-    if (route.hasOwnProperty('before') && (typeof route.before === 'function')) {
-      // Handle the route.
-      return route.before(req, res, function(err) {
-        if (err) {
-          throw err;
-        }
-
-        return Q();
-      });
-    }
-
-    return Q();
-  };
-
-  /**
-   * Check if the current route has a after fn defined, if so, execute it and proceed.
-   *
-   * @param {object} route
-   *   The route object.
-   * @param {object} req
-   *   The express request object.
-   * @param {object} res
-   *   The express response object.
-   *
-   * @returns {Promise}
-   */
-  var after = function after(route, req, res) {
-    debug('After:');
-    debug(route);
-
-    // Let the route also define its own handler.
-    if (
-      route.hasOwnProperty('after') &&
-      (typeof route.after === 'function')
-    ) {
-      // Handle the route.
-      return route.after(req, res, function(err, result) {
-        result = result || res.result;
-        if (err) {
-          throw err;
-        }
-
-        // Send the result.
-        debug('Result:');
-        debug(result);
-        return res.status(result.status).send(result);
-      });
-    }
-
-    // Send the result.
-    debug(res.result);
-    return res.status(res.result.status).send(res.result);
-  };
-
   return {
     connect: connect,
     request: request,
-    query: query,
-    before: before,
-    after: after
+    query: query
   };
 };
